pyproject.toml
[tool.poetry]
name = "djdt-permissions"
version = "0.1.0"
description = ""
authors = ["Alain Baudhuin <alain.baudhuin@skynet.be>"]
readme = "README.md"
packages = [{include = "djdt_permissions"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0"  # TODO MATCH DjDT dependencies.
django-debug-toolbar = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
